add checkstyle
