+++ src/main/java/command/DeadlineCommand.java
@@ -1,16 +1,24 @@
-package Command;
-import List.TaskList;
-import Ui.Ui;
-import FileStorage.FileStorage;
-import DukeException.DukeException;
-import Tasks.Deadline;
+package command;
+
+import dukeexception.DukeException;
+import filestorage.FileStorage;
+import list.TaskList;
+import tasks.Deadline;
+import ui.Ui;
 
 /**
  * A class that is part of the command, for this in particular this will store a task that has a deadline.
  */
-public class DeadlineCommand extends Command{
-    private String task;
-    private String deadDate;
+public class DeadlineCommand extends Command {
+    private final String task;
+    private final String deadDate;
+
+    /**
+     * A constructor method to call for the method.
+     *
+     * @param task the task that is input by the user.
+     * @param deadDate the deadline of the task.
+     */
     public DeadlineCommand(String task, String deadDate) {
         this.task = task;
         this.deadDate = deadDate;
