+++ src/main/java/command/MarkCommand.java
@@ -1,22 +1,24 @@
-package Command;
+package command;
 
-import List.TaskList;
-import Ui.Ui;
-import FileStorage.FileStorage;
-import DukeException.DukeException;
-import Tasks.Task;
+import dukeexception.DukeException;
+import filestorage.FileStorage;
+import list.TaskList;
+import tasks.Task;
+import ui.Ui;
+
 
 /**
- * A class that will call for the unmark command
+ * A class that will mark the task as done.
  */
-public class UnmarkCommand extends Command{
-    private int index;
-    public UnmarkCommand(int index) {
+public class MarkCommand extends Command {
+
+    private final int index;
+    public MarkCommand(int index) {
         this.index = index;
     }
 
     /**
-     * A method that will mark the task as unMarked.
+     * A method that will excute the mark command depending on the index by user .
      *
      * @param tasklist contains all the past few tasks excuted.
      * @param ui contains the user interface that will be shown to the user depending on the inputs.
@@ -24,8 +26,8 @@ public class UnmarkCommand extends Command{
      * @throws DukeException If user inputs is invalid.
      */
     public void excute(TaskList tasklist, Ui ui, FileStorage fileStorage) throws DukeException {
-        Task task = tasklist.unmarkTask(this.index);
+        Task task = tasklist.markTask(this.index);
         fileStorage.write(tasklist);
-        ui.showUnMarkedTask(task);
+        ui.showMarkedTask(task);
     }
 }
