+++ src/main/java/command/UnmarkCommand.java
@@ -1,24 +1,22 @@
-package Command;
+package command;
 
-import List.TaskList;
-import Ui.Ui;
-import FileStorage.FileStorage;
-import DukeException.DukeException;
-import Tasks.Task;
+import dukeexception.DukeException;
+import filestorage.FileStorage;
+import list.TaskList;
+import tasks.Task;
+import ui.Ui;
 
 /**
- * A class that will mark the task as done.
+ * A class that will call for the unmark command
  */
-public class MarkCommand extends Command{
-
-    private int index;
-    public MarkCommand(int index) {
+public class UnmarkCommand extends Command {
+    private final int index;
+    public UnmarkCommand(int index) {
         this.index = index;
     }
 
     /**
-     * A method that will excute the mark command depending on the index by user .
-     *
+     * A method that will mark the task as unMarked.
      *
      * @param tasklist contains all the past few tasks excuted.
      * @param ui contains the user interface that will be shown to the user depending on the inputs.
@@ -26,8 +24,8 @@ public class MarkCommand extends Command{
      * @throws DukeException If user inputs is invalid.
      */
     public void excute(TaskList tasklist, Ui ui, FileStorage fileStorage) throws DukeException {
-        Task task = tasklist.markTask(this.index);
+        Task task = tasklist.unmarkTask(this.index);
         fileStorage.write(tasklist);
-        ui.showMarkedTask(task);
+        ui.showUnMarkedTask(task);
     }
 }
