+++ src/main/java/tasks/Todo.java
@@ -1,0 +1,15 @@
+package tasks;
+
+/**
+ * A child class to Task, this in particular is for todo tasks.
+ */
+public class Todo extends Task {
+    /**
+     * A constructor method for Todo.
+     * @param name the name of the task.
+     */
+    public Todo(String name) {
+        super(name);
+        super.setType("T");
+    }
+}
